# crates/client/openapi.yaml
openapi: 3.1.0
info:
  title: Silicon Remind API
  version: 0.1.0
  description: Durable Silicon reminders with organization-wide read access, owner-Silicon mutations,
    IAM application sessions and isolated test environments. Test requests use X-Remind-Test-Key plus
    a test IAM identity; they never fall through to production.
servers:
- url: https://backend.remind.teamofsilicons.com/api/v1
security:
- bearerAuth: []
...
paths:
  /schedules:
    get:
      tags:
      - Schedules
      operationId: listSchedules
      summary: List schedules visible to the current actor
      description: Current or archived reminders in the authenticated organization, filtered before cursor
        pagination. Carbons are read-only; only an owning Silicon may mutate reminders.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - name: silicon_id
        in: query
        schema:
          $ref: '#/components/schemas/SiliconGlobalId'
      - name: section
        in: query
        description: Product-facing reminder section; current is the default.
        schema:
          allOf:
          - $ref: '#/components/schemas/ScheduleSection'
          default: current
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/ScheduleStatus'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/TestKey'
      responses:
        '200':
          description: Schedules
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Schedule'
                  next_cursor:
                    type:
                    - string
                    - 'null'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags:
      - Schedules
      operationId: updateScheduleStatuses
      summary: Atomically pause or resume owned current schedules
      description: Applies one active or paused status to every requested schedule in the supplied order.
        The authenticated Silicon must own every schedule, and every schedule must still be current. Validation
        is all-or-nothing; failures leave the complete batch unchanged. The response contains one compact
        status result per requested schedule in request order.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/TestKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleStatusBatchUpdate'
      responses:
        '200':
          description: Schedule status results in request order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleStatusBatchUpdateResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags:
      - Schedules
      operationId: createSchedule
      summary: Create a one-time or recurring schedule for the authenticated Silicon
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/TestKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleCreate'
      responses:
        '201':
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '409':
          $ref: '#/components/responses/WebhookNotConfigured'
        default:
          $ref: '#/components/responses/Error'
  /schedules/{schedule_id}:
    parameters:
    - $ref: '#/components/parameters/OrgId'
    - $ref: '#/components/parameters/ScheduleId'
    get:
      tags:
      - Schedules
      operationId: getSchedule
      summary: Get an IAM-visible schedule
      description: Inaccessible identifiers return the same not-found response as absent identifiers.
      responses:
        '200':
          description: Schedule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        default:
          $ref: '#/components/responses/Error'
      parameters:
      - $ref: '#/components/parameters/TestKey'
    patch:
      tags:
      - Schedules
      operationId: updateSchedule
      summary: Update a schedule owned by the authenticated Silicon
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/TestKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              additionalProperties: false
              properties:
                text:
                  type: string
                  minLength: 1
                timezone:
                  type: string
                  description: IANA time-zone identifier
                kind:
                  $ref: '#/components/schemas/ScheduleKind'
                cron:
                  type: string
                  description: Five-field Linux cron expression
                status:
                  $ref: '#/components/schemas/MutableScheduleStatus'
      responses:
        '200':
          description: Updated schedule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags:
      - Schedules
      operationId: archiveSchedule
      summary: Archive a schedule owned by the authenticated Silicon for 45 days
      responses:
        '204':
          description: Archived, or already archived
        default:
          $ref: '#/components/responses/Error'
      parameters:
      - $ref: '#/components/parameters/TestKey'
  /schedules/{schedule_id}/executions:
    get:
      tags:
      - Executions
      operationId: listScheduleExecutions
      summary: List execution history for an IAM-visible schedule
      description: The parent schedule's IAM-projected owner visibility is enforced before pagination.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/ScheduleId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/TestKey'
      responses:
        '200':
          description: Execution history
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Execution'
                  next_cursor:
                    type:
                    - string
                    - 'null'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: IAM Remind-audience application token
    testKeyAuth:
      type: apiKey
      in: header
      name: X-Remind-Test-Key
      description: 32-character Remind root key. Selects an isolated test environment; ordinary actions
        still require a test IAM access token.
  parameters:
    OrgId:
      name: X-Org-ID
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/OrgId'
    ScheduleId:
      name: schedule_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 255
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    TestKey:
      name: X-Remind-Test-Key
      in: header
      required: false
      description: Optional sandbox selector. Invalid, expired or revoked keys fail closed. Environment
        management routes reject this header.
      schema:
        type: string
        pattern: ^[a-zA-Z0-9]{32}$
  responses:
    WebhookNotConfigured:
      description: The authenticated Silicon has no active Hook destination.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: webhook_not_configured
              message: Set the webhook url first.
              request_id: 0199d8d5-2f62-7eb7-81c5-1e456778f8de
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    OrgId:
      type: string
      pattern: ^[a-z0-9_-]{3,50}$
    SiliconGlobalId:
      type: string
      pattern: ^[a-z0-9_-]{3,50}:[a-z0-9_-]{3,50}$
    ScheduleStatus:
      type: string
      enum:
      - active
      - paused
      - completed
    ScheduleSection:
      type: string
      enum:
      - current
      - archived
    MutableScheduleStatus:
      type: string
      enum:
      - active
      - paused
    ScheduleStatusBatchUpdate:
      type: object
      required:
      - schedule_ids
      - status
      additionalProperties: false
      properties:
        schedule_ids:
          type: array
          minItems: 1
          maxItems: 100
          uniqueItems: true
          items:
            type: string
            format: uuid
        status:
          $ref: '#/components/schemas/MutableScheduleStatus'
    ScheduleStatusBatchUpdateResponse:
      type: object
      required:
      - items
      additionalProperties: false
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/ScheduleStatusResult'
    ScheduleStatusResult:
      type: object
      required:
      - id
      - status
      - next_run_at
      - updated_at
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/MutableScheduleStatus'
        next_run_at:
          type:
          - string
          - 'null'
          format: date-time
        updated_at:
          type: string
          format: date-time
    ScheduleKind:
      type: string
      enum:
      - one_time
      - recurring
    ScheduleCreate:
      type: object
      required:
      - text
      - kind
      - cron
      additionalProperties: false
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 100000
        kind:
          $ref: '#/components/schemas/ScheduleKind'
        cron:
          type: string
          description: Five-field Linux cron expression
        timezone:
          type: string
          default: UTC
          description: IANA time-zone identifier
          examples:
          - Asia/Kolkata
    Schedule:
      allOf:
      - $ref: '#/components/schemas/ScheduleCreate'
      - type: object
        required:
        - id
        - org_id
        - silicon_id
        - timezone
        - status
        - section
        - next_run_at
        - archived_at
        - purge_after
        - created_at
        - updated_at
        properties:
          id:
            type: string
            format: uuid
          org_id:
            $ref: '#/components/schemas/OrgId'
          silicon_id:
            $ref: '#/components/schemas/SiliconGlobalId'
          status:
            $ref: '#/components/schemas/ScheduleStatus'
          section:
            $ref: '#/components/schemas/ScheduleSection'
          next_run_at:
            type:
            - string
            - 'null'
            format: date-time
          archived_at:
            type:
            - string
            - 'null'
            format: date-time
          purge_after:
            type:
            - string
            - 'null'
            format: date-time
          created_at:
            type: string
            format: date-time
          updated_at:
            type: string
            format: date-time
    Execution:
      type: object
      required:
      - id
      - schedule_id
      - scheduled_for
      - status
      properties:
        id:
          type: string
          format: uuid
          description: Stable idempotency identifier for this occurrence
        schedule_id:
          type: string
          format: uuid
        scheduled_for:
          type: string
          format: date-time
        attempted_at:
          type:
          - string
          - 'null'
          format: date-time
        delivered_at:
          type:
          - string
          - 'null'
          format: date-time
        status:
          type: string
          enum:
          - pending
          - delivered
          - retrying
          - failed
        hook_event_id:
          type:
          - string
          - 'null'
          format: uuid
        failure_reason:
          type:
          - string
          - 'null'
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - request_id
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
...